Renames folder missed by PR#1068
Closes-bug: APPLINK-30334
