+++ src/components/application_manager/include/application_manager/policies/delegates/app_permission_delegate.h
@@ -1,5 +1,5 @@
 #ifdef EXTERNAL_PROPRIETARY
-#include "application_manager/policies/premium/delegates/app_permission_delegate.h"
+#include "application_manager/policies/external/delegates/app_permission_delegate.h"
 #else
 #include "application_manager/policies/regular/delegates/app_permission_delegate.h"
 #endif
