+++ src/components/application_manager/include/application_manager/policies/delegates/statistics_delegate.h
@@ -1,5 +1,5 @@
 #ifdef EXTERNAL_PROPRIETARY
-#include "application_manager/policies/premium/delegates/statistics_delegate.h"
+#include "application_manager/policies/external/delegates/statistics_delegate.h"
 #else
 #include "application_manager/policies/regular/delegates/statistics_delegate.h"
 #endif
