+++ src/components/application_manager/include/application_manager/policies/policy_event_observer.h
@@ -1,5 +1,5 @@
 #ifdef EXTERNAL_PROPRIETARY
-#include "application_manager/policies/premium/policy_event_observer.h"
+#include "application_manager/policies/external/policy_event_observer.h"
 #else
 #include "application_manager/policies/regular/policy_event_observer.h"
 #endif
