+++ src/components/application_manager/include/application_manager/policies/policy_retry_sequence.h
@@ -1,5 +1,5 @@
 #ifdef EXTERNAL_PROPRIETARY
-#error "policy_handler_observer.h is not available in policy premium"
+#error "policy_handler_observer.h is not available in policy external"
 #else
 #include "application_manager/policies/regular/policy_handler_observer.h"
 #endif
