+++ src/components/application_manager/include/application_manager/policies/pt_exchange_handler_impl.h
@@ -1,0 +1,6 @@
+#ifdef EXTERNAL_PROPRIETARY
+#error "pt_exchange_handler_impl.h is not available in policy external"
+#else
+#include "application_manager/policies/regular/pt_exchange_handler_impl.h"
+#endif
+
