+++ src/components/application_manager/src/policies/policy_handler.cc
@@ -741,7 +741,7 @@ void PolicyHandler::OnVehicleDataUpdated(
   }
 #else
   LOG4CXX_DEBUG(logger_,
-                "This functionality is not available for not premium policy");
+                "This functionality is not available for not external policy");
 #endif
 }
